refactor: remove MemoryEventHandler from memory_client and integrate it into memory_server
