+++ src/c_two/rpc/memory/memory_server.py
@@ -7,15 +7,30 @@ import tempfile
 import threading
 from pathlib import Path
 from watchdog.observers import Observer
+from watchdog.events import FileSystemEventHandler
 
 from ..util.wait import wait
 from ..base import BaseServer
 from ..event import Event, EventTag, EventQueue
-from .event_handler import MemoryEventHandler
 
 logger = logging.getLogger(__name__)
 
 MAXIMUM_WAIT_TIMEOUT = 1
+
+class MemoryEventHandler(FileSystemEventHandler):
+    def __init__(self, file_pattern: str, file_extension: str = '.mem'):
+        self.file_received = False
+        self.file_pattern = file_pattern
+        self.file_extension = file_extension
+        self.condition = threading.Condition()
+    
+    def on_moved(self, event):
+        if not event.is_directory:
+            dest_filename = os.path.basename(event.dest_path)
+            if dest_filename.startswith(self.file_pattern) and dest_filename.endswith(self.file_extension):
+                with self.condition:
+                    self.file_received = True
+                    self.condition.notify_all() # wake up all waiting threads
 
 class MemoryServerState(enum.Enum):
     STOPPED = 'stopped'
